MOVEONLY: Move logging code from util.{h,cpp} to new files.
backports https://github.com/bitcoin/bitcoin/commit/b77b6e23457dcbf160a3d586eca9df7a8d0bb351
